diag model separated from ui. fix for diag for class ctxt menu not getting refilled
