+++ sveditor/plugins/net.sf.sveditor.core/src/net/sf/sveditor/core/diagrams/IDiagModelFactory.java
@@ -9,7 +9,8 @@
  *     Armond Paiva - initial contributor 
  ****************************************************************************/
 
-package net.sf.sveditor.ui.views.diagram;
+package net.sf.sveditor.core.diagrams;
+
 
 public interface IDiagModelFactory {
