+++ sveditor/plugins/net.sf.sveditor.ui/src/net/sf/sveditor/ui/views/diagram/figures/UMLClassFigure.java
@@ -11,7 +11,8 @@
  * 
  ******************************************************************************/
 
-package net.sf.sveditor.ui.views.diagram;
+package net.sf.sveditor.ui.views.diagram.figures;
+
 
 import org.eclipse.draw2d.ColorConstants;
 import org.eclipse.draw2d.Figure;
